requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
clickhouse-driver==0.2.3
clickhouse-sqlalchemy==0.2.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
pydantic==1.9.1
pyhumps==3.5.3
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
psycopg2~=2.9
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.17.6
